+++ app/src/test/java/app/itochan/aes67/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package app.itochan.aes67mic
+package app.itochan.aes67
 
 import org.junit.Test
